Tescik

+++ library/src/test/java/net/sf/txt2srt/ConverterTest.java
@@ -6,7 +6,7 @@ import java.io.File;
 import java.io.FileNotFoundException;
 import java.io.IOException;
 
-import org.fest.assertions.Assertions;
+import org.junit.After;
 import org.junit.Before;
 import org.junit.Test;
 
@@ -15,27 +15,33 @@ public class ConverterTest {
 	private static final int DURATION = 0;
 	private static final int FRAMERATE = 25;
 	private static final String ENCODING_CP1250 = "cp1250";
-	private static final String SRC_TEST_RESOURCES_SAMPLE_TXT = "src/test/resources/sample.txt";
-	private static final String SRC_TEST_RESOURCES_SAMPLE_SRT = "src/test/resources/sample.srt";
-	private static final String TARGET_SAMPLE_SRT = "target/sample.srt";
+	private static final String INPUT_SAMPLE_SUBTITLE_MPL2 = "src/test/resources/sample.txt";
+	private static final String CONVERTED_SAMPLE_SUBTITLE_SRT = "src/test/resources/sample.srt";
+	private static final String CONVERTED_SAMPLE_SUBTITLE_SRT_TARGET = "target/sample.srt";
+
 	Converter conv;
+
+	private File resultFile;
 
 	@Before
 	public void setup() {
+		resultFile = new File(CONVERTED_SAMPLE_SUBTITLE_SRT);
+		cleanup();
+		assertThat(resultFile.isFile()).isFalse();
 		conv = new Converter();
 	}
 
-	@Test
-	public void clean() {
-		File resultFile = new File(TARGET_SAMPLE_SRT);
-		resultFile.delete();
-		assertThat(!resultFile.exists()).isTrue();
+	@After
+	public void cleanup() {
+		if (resultFile != null && resultFile.exists()) {
+			resultFile.delete();
+		}
 	}
 
 	@Test
 	public void shouldReadTypeFromSpecifiedSubtitleFile()
 			throws FileNotFoundException, IOException {
-		conv.readSubtitle(SRC_TEST_RESOURCES_SAMPLE_TXT, ENCODING_CP1250,
+		conv.readSubtitle(INPUT_SAMPLE_SUBTITLE_MPL2, ENCODING_CP1250,
 				FRAMERATE, DURATION);
 
 		assertThat(conv.getSubtitleType()).isNotEmpty();
@@ -43,47 +49,41 @@ public class ConverterTest {
 
 	}
 
-	public void shouldWriteSubtitle()
-			throws FileNotFoundException, IOException {
-		conv.readSubtitle(SRC_TEST_RESOURCES_SAMPLE_TXT, ENCODING_CP1250,
+	public void shouldWriteSubtitle() throws FileNotFoundException, IOException {
+
+		conv.readSubtitle(INPUT_SAMPLE_SUBTITLE_MPL2, ENCODING_CP1250,
 				FRAMERATE, DURATION);
 
 		String result = conv.writeSubtitle();
-		File resultFile = new File(result);
-		
-		assertThat(resultFile.exists() && resultFile.isFile()).isTrue();
 
-		// cleanup
-		resultFile.delete();
+		assertThat(result).isEqualTo(resultFile.getPath());
+		assertThat(resultFile.exists() && resultFile.isFile()).isTrue();
 
 	}
 
 	@Test
 	public void shouldCreateSrtSubtitle() throws IOException {
 
-		conv.convert(SRC_TEST_RESOURCES_SAMPLE_TXT, TARGET_SAMPLE_SRT,
-				ENCODING_CP1250, FRAMERATE, 0);
+		conv.convert(INPUT_SAMPLE_SUBTITLE_MPL2,
+				CONVERTED_SAMPLE_SUBTITLE_SRT_TARGET, ENCODING_CP1250,
+				FRAMERATE, 0);
 
-		File resultFile = new File(TARGET_SAMPLE_SRT);
-		// File.
-		assertThat(resultFile.isFile()).isTrue();
+		resultFile = new File(CONVERTED_SAMPLE_SUBTITLE_SRT_TARGET);
+
 		assertThat(resultFile.exists() && resultFile.isFile()).isTrue();
 	}
 
 	@Test
 	public void shouldCreateSrtSubtitleInThisSameDir() throws IOException {
-		File resultFile = new File(SRC_TEST_RESOURCES_SAMPLE_SRT);
-		assertThat(resultFile.isFile()).isFalse();
-		
+
 		conv = new Converter();
 
-		String res = conv.convert(SRC_TEST_RESOURCES_SAMPLE_TXT,
+		String result = conv.convert(INPUT_SAMPLE_SUBTITLE_MPL2,
 				ENCODING_CP1250, FRAMERATE, 0);
 
-		System.out.println("Result is " + res);
+		System.out.println("Result is " + result);
+		assertThat(result).isEqualTo(resultFile.getPath());
 		assertThat(resultFile.exists() && resultFile.isFile()).isTrue();
 
-		// cleanup
-		resultFile.delete();
 	}
 }
